+++ CHANGELOG.md
@@ -18,6 +18,7 @@ any parts of the framework not mentioned in the documentation should generally b
 
 * Added support to overwrite serializer methods in customized schema class
 * Adjusted some still old formatted strings to f-strings.
+* Replaced `OrderedDict` with `dict` which is also ordered since Python 3.7.
 
 ### Fixed