+++ CHANGELOG.md
@@ -13,6 +13,7 @@ any parts of the framework not mentioned in the documentation should generally b
 ### Added
 
 * Added support for Python 3.11.
+* Added support for Django 4.2.
 
 ### Changed